minor indentation, spacing, etc fixes
