+++ index.html
@@ -1,19 +1,15 @@
 <!DOCTYPE html>
-<html lang = "en">
-
+<html lang="en">
    <head>
-      <meta charset = "UTF-8">
+      <meta charset="UTF-8">
       <title>React App</title>
-	  <meta name="viewport" content="width=device-width, initial-scale=1">
-	  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-	  <link rel="stylesheet" href="main.css">
+      <meta name="viewport" content="width=device-width, initial-scale=1">
+      <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+      <link rel="stylesheet" href="main.css">
    </head>
 
    <body>
-      <div id = "app"></div> 
-	  <script src="https://unpkg.com/react@15/dist/react-with-addons.js"></script>
-	  <script src="https://unpkg.com/react-router/umd/ReactRouter.min.js"></script>
-      <script src = "index.js"></script>
+      <div id="app"></div> 
+      <script src="index.js"></script>
    </body>
-
 </html>
