+++ modules/SliderContentList.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+
+import ContentSliderNavigation from './ContentSliderNavigation.jsx';
+import Slider from './Slider.jsx';
+
+export default class SliderContentList extends React.Component {
+  render() {
+    let current = this.props.currentSlide;
+
+    return (
+      <div>
+        <div className="accordion-content" style={{'maxHeight': this.props.isCollapsed ? '0px' : '1000px' }} >  
+          <ul className="sliders">
+            {
+              this.props.datalist.map(function(item, index) {
+                let active = current.key === item.key;
+                return (
+                  <li key={item.key} className={(active ? 'current' : '')}>
+                    <Slider data={item}  />
+                  </li>
+                );
+              })
+            }
+          </ul>
+
+          <ContentSliderNavigation 
+            nextButtonClick={this.props.nextButtonClick} backButtonClick={this.props.backButtonClick} 
+            currentSlide={this.props.currentSlide} 
+            datalist={this.props.datalist}
+          />
+
+        </div>
+      </div>
+    );
+  }
+}
