+++ package.json
@@ -14,19 +14,17 @@
     "react-router": "^2.0.0"
   },
   "devDependencies": {
-"css-loader": "^0.18.0",
+    "css-loader": "^0.18.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "node-sass": "^3.13.0",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
-	 "http-server": "^0.8.5",
+    "http-server": "^0.8.5",
     "babel-core": "^6.5.1",
     "babel-loader": "^6.2.2",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-react": "^6.5.0",
-   
     "webpack": "^1.12.13",
     "webpack-dev-server": "^1.14.1"
-    
   }
 }
